refactor(components/ui/pagination): keep default patterns
