+++ src/presentation/components/ui/pagination/content.tsx
@@ -1,16 +1,15 @@
-import React from 'react'
+import { forwardRef, ComponentProps } from 'react'
 
 import { cn } from '@/main/utils'
 
-export const Content = React.forwardRef<
-	HTMLUListElement,
-	React.ComponentProps<'ul'>
->(({ className, ...props }, ref) => (
-	<ul
-		ref={ref}
-		className={cn('flex flex-row items-center gap-1', className)}
-		{...props}
-	/>
-))
+export const Content = forwardRef<HTMLUListElement, ComponentProps<'ul'>>(
+	({ className, ...props }, ref) => (
+		<ul
+			ref={ref}
+			className={cn('flex flex-row items-center gap-1', className)}
+			{...props}
+		/>
+	)
+)
 
 Content.displayName = 'PaginationContent'
