+++ src/presentation/components/ui/pagination/index.ts
@@ -3,8 +3,8 @@ import { Ellipsis } from './ellipsis'
 import { Item } from './item'
 import { Link } from './link'
 import { Next } from './next'
-import { Root } from './pagination'
 import { Previous } from './previous'
+import { Root } from './root'
 
 export const Pagination = {
 	Root,
